requirements.txt
faiss-cpu>=1.8.0
numpy>=1.26.4
truststore>=0.8.0
PyPDF2>=3.0.1
python-dateutil>=2.9.0
streamlit>=1.36.0
pydantic>=2.7.4
rank-bm25>=0.2.2
sentence-transformers>=2.7.0
pyyaml>=6.0.1
FlagEmbedding>=1.2.10
loguru>=0.7.2
python-dotenv>=1.0.1
